WIP: sync repo data from source controller
